refactor(shared): 権限の「知識」を sandbox から shared へ集約
capability カタログ・危険度・permissionPolicy は Worker/DOM 非依存の純粋な
共有知識であり、react/frontend/build-workers/docs 生成が参照する。sandbox
（Worker 隔離パッケージ）ではなく shared に置くことで、backend/CLI 等も
Worker コード抜きで参照できる。実行時の enforcement（capabilityGate）は
host（sandbox）に残置。新パッケージは作らず既存 shared に集約（過分離を避ける）。

- packages/shared/src/mod/capability.ts, permission.ts へ移動（+ テスト）。
- CMD_TO_HANDLER（HostHandlers 依存の host 固有マップ）は sandbox の
  commandHandlers.ts へ分離。
- shared/index.ts に mod/capability・mod/permission を re-export。
- sandbox/index.ts から権限 re-export を削除（capabilityGate のみ残す）。
- react は権限を @ubichill/shared から再エクスポート（診断は sandbox のまま）。
- gen-capability-docs.mjs / build-workers.mjs / docs のパス参照を shared に更新。
