+++ packages/react/src/components/PermissionContext.test.tsx
@@ -1,6 +1,6 @@
 // @vitest-environment jsdom
 import { act, renderHook } from '@testing-library/react';
-import type { PermissionPolicy } from '@ubichill/sandbox';
+import type { PermissionPolicy } from '@ubichill/shared';
 import type React from 'react';
 import { describe, expect, it, vi } from 'vitest';
 import { PermissionProvider, useUbiPermissions } from './PermissionContext';
